pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aat-devqa"
version = "0.2.0"
description = "AI Auto Tester — AI-powered DevQA Loop Orchestrator"
readme = "README.md"
license = "AGPL-3.0-only"
requires-python = ">=3.11"
authors = [{ name = "AAT Contributors" }]
keywords = ["testing", "automation", "ai", "devqa"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
]

dependencies = [
    # CLI
    "typer[all]>=0.9,<1.0",
    # Config
    "pydantic>=2.5,<3.0",
    "pydantic-settings>=2.1,<3.0",
    "pyyaml>=6.0,<7.0",
    # Engine
    "playwright>=1.40,<2.0",
    # Image Matching
    "opencv-python-headless>=4.8,<5.0",
    "numpy>=1.24,<3.0",
    "pillow>=10.0,<12.0",
    # OCR
    "pytesseract>=0.3.10,<1.0",
    # AI
    "anthropic>=0.40,<1.0",
    # Reporter
    "jinja2>=3.1,<4.0",
    # Misc
    "rich>=13.0,<14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-asyncio>=0.23,<1.0",
    "pytest-cov>=4.0,<6.0",
    "ruff>=0.4,<1.0",
    "mypy>=1.8,<2.0",
    "pre-commit>=3.6,<4.0",
    "pytest-mock>=3.12,<4.0",
    "types-PyYAML>=6.0,<7.0",
]

[project.scripts]
aat = "aat.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/aat"]

[tool.ruff]
target-version = "py311"
line-length = 99
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "TCH"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src/aat"]
omit = ["*/tests/*"]
